requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
tqdm==4.66.2
pypdf==4.1.0
python-docx==1.1.0
gdown==4.7.3
